refactor(Dropdown): Isolated function normalizeChars in shared folder
